Reorganise services
