+++ src/services/ticket/ticket.service.ts
@@ -1,276 +1,27 @@
-import { Ticket, TicketStatus, TicketMessage } from '../../model/ticket';
-import { User } from '../../model/user';
-import { Line } from '../../model/line';
-import { ApiBase } from '../api-base/api-base';
 import {
-  extractId,
-  extractIdToNumber,
-  IdOrObject,
-} from '../../util/id-or-object';
-
-/**
- * Represents a collection of search criteria for TicketService.count().
- *
- * All of the criteria that have been included must match for tickets to be included.
- *
- * For example, the query `{ line: [123, 34], status: ['SERVED'] }` will retrieve a count of
- * tickets that were marked served in lines 123 and 34.
- */
-interface TicketCountCriteria {
-  /**
-   * An array of lines to search tickets from, represented as an array of line IDs.
-   *
-   * For example:
-   * `line: [123, 234, 456]`
-   * `line: '123,234,456'`
-   */
-  line?: string | Array<number> | number;
-  /**
-   * The location ID to search tickets from.
-   *
-   * For example: `location: 4`
-   */
-  location?: number;
-  /**
-   * The ticket statuses to include (for example new, called, served, no-shows)
-   *
-   * For example, `status: ['NEW', 'CALLED', 'SERVED']`
-   */
-  status?: Array<TicketStatus> | string;
-  /**
-   * The caller of the ticket, for finding tickets that were called by a particular user.
-   * Tickets that are NEW have not been called - they won't be included in results with a caller
-   * filter.
-   *
-   * For example, `caller: 13410`
-   */
-  caller?: IdOrObject<User>;
-  /**
-   * The minimum creation date of the ticket, using either a Unix timestamp or an ISO8601 date.
-   * If minCreated is specified, search results will only include tickets created after minCreated.
-   *
-   * For example, `minCreated: '2017-09-02T10:12:10Z'` or `minCreated: 1504348504`
-   */
-  minCreated?: string | number;
-  /**
-   * The maximum creation date of the ticket.
-   * If maxCreated is specified, search results will only include tickets created before maxCreated.
-   *
-   * For example, `maxCreated: '2017-09-02T10:12:10Z'` or `maxCreated: 1504348504`
-   */
-  maxCreated?: string | number;
-  /**
-   * The minimum called date of the ticket.
-   * If minCalled is specified, search results will only include tickets called after minCalled.
-   *
-   * For example, `minCalled: '2017-09-02T10:12:10Z'` or `minCalled: 1504348504`
-   */
-  minCalled?: string | number;
-  /**
-   * The maximum called date of the ticket.
-   * If maxCalled is specified, search results will only include tickets called before maxCalled.
-   *
-   * For example, `maxCalled: '2017-09-02T10:12:10Z'` or `maxCalled: 1504348504`
-   */
-  maxCalled?: string | number;
-}
-
-/**
- * Represents a collection of search criteria for TicketService.search().
- * All of these criteria need to apply in order for tickets to be included.
- *
- * For example, the query `{ line: [123, 34], status: ['SERVED'] }` will retrieve all tickets that
- * were marked served in lines 123 and 34.
- *
- * The query `{ status: ['CALLED'], caller: 11111, limit: 1 }` will retrieve the clerk 11111's
- * currently called ticket.
- */
-interface TicketSearchCriteria {
-  /**
-   * An array of lines to search tickets from, represented as an array of line IDs.
-   *
-   * For example: `line: [123, 234, 456]`, or `line: 1234`, or `line: '1234,2434,5555'`
-   */
-  line?: string | Array<number> | number;
-  /**
-   * The location ID to search tickets from.
-   *
-   * For example: `location: 4`
-   */
-  location?: number;
-  /**
-   * The ticket statuses to include (for example new, called, served, no-shows)
-   *
-   * For example, `status: ['NEW', 'CALLED', 'SERVED']`
-   */
-  status?: Array<TicketStatus> | string;
-  /**
-   * The caller of the ticket, for finding tickets that were called by a particular user.
-   * Tickets that are NEW have not been called - they won't be included in results with a caller
-   * filter.
-   *
-   * For example, `caller: 13410`
-   */
-  caller?: IdOrObject<User>;
-  /**
-   * The minimum creation date of the ticket, using either a Unix timestamp or an ISO8601 date.
-   * If minCreated is specified, search results will only include tickets created after minCreated.
-   *
-   * For example, `minCreated: '2017-09-02T10:12:10Z'` or `minCreated: 1504348504`
-   */
-  minCreated?: string | number;
-  /**
-   * The maximum creation date of the ticket.
-   * If maxCreated is specified, search results will only include tickets created before maxCreated.
-   *
-   * For example, `maxCreated: '2017-09-02T10:12:10Z'` or `maxCreated: 1504348504`
-   */
-  maxCreated?: string | number;
-  /**
-   * The minimum called date of the ticket.
-   * If minCalled is specified, search results will only include tickets called after minCalled.
-   *
-   * For example, `minCalled: '2017-09-02T10:12:10Z'` or `minCalled: 1504348504`
-   */
-  minCalled?: string | number;
-  /**
-   * The maximum called date of the ticket.
-   * If maxCalled is specified, search results will only include tickets called before maxCalled.
-   *
-   * For example, `maxCalled: '2017-09-02T10:12:10Z'` or `maxCalled: 1504348504`
-   */
-  maxCalled?: string | number;
-  /**
-   * The maximum number of results retrieved.
-   * The number has to be between 1 and 10000, and by default up to 1000 results will be retrieved.
-   *
-   * For example, `limit: 50`
-   */
-  limit?: number;
-  /**
-   * The order of the returned tickets.
-   *
-   * Tickets can be sorted by the following keys: id, number, created, called.
-   *
-   * Additionally, you may specify ascending or descending order by adding 'ASC' or 'DESC' after
-   * the key, separated by spaces.
-   *
-   * If this property is omitted, the default ordering is 'id ASC'.
-   *
-   * For example, `order: 'id ASC'`, or `order: 'created DESC'`
-   */
-  order?: string;
-
-  /**
-   * This property decides whether to include additional information along with the tickets.
-   *
-   * It is possible to add the ticket's **messages** and **interactions** to each ticket in the
-   * response, by specifying the responseScope.
-   *
-   * If omitted, the ticket will not include the message list or interaction list.
-   *
-   * If the value is 'MESSAGES', the ticket's message list is attached to each ticket. Tickets
-   * without messages have an empty array.
-   *
-   * If the value is 'INTERACTIONS', the ticket's interaction list is attached to each ticket.
-   * Tickets without interactions have an empty array.
-   *
-   * For example:
-   *
-   * responseScope: 'MESSAGES'
-   * responseScope: ['MESSAGES', 'INTERACTIONS']
-   * responseScope: 'MESSAGES,INTERACTIONS'
-   */
-  responseScope?:
-    | string
-    | Array<'MESSAGES' | 'INTERACTIONS'>
-    | 'MESSAGES'
-    | 'INTERACTIONS';
-}
-
-/**
- * Represents a position in the queue where the ticket should go when returning it to the queue.
- * They should either be the first in line or the last in line.
- */
-type DesiredQueuePosition = 'FIRST' | 'LAST';
-
-/** This error is thrown when a Line ID is not passed to the API method, or when its type is not
- *  a number.
- *  @hidden */
-export const ERROR_NO_LINE_ID: string = 'Line ID missing from arguments.';
-/**
- * This error is thrown when the line is not a number (for the line ID) or a valid Line object.
- * @hidden
- */
-export const ERROR_INVALID_LINE: string =
-  'Line is not a number or Line object.';
-
-/** This error is thrown when the Ticket ID is not passed to the API method, or when its type is
- *  not a number.
- *  @hidden */
-export const ERROR_NO_TICKET_ID: string = 'Ticket ID missing from arguments.';
-/** @hidden */
-export const ERROR_INVALID_TICKET: string =
-  'Ticket is not a number or Ticket object.';
-/** This error is thrown when the Ticket Changes object is not passed to TicketService.edit.
- * @hidden */
-export const ERROR_NO_TICKET_CHANGES: string =
-  'Ticket changes missing from arguments.';
-
-/** This error is thrown when a user is not passed into API methods that take a user.
- * @hidden */
-export const ERROR_NO_USER: string = 'User missing from arguments.';
-/** @hidden */
-export const ERROR_INVALID_USER: string =
-  'User is not a number or User object.';
-/** This error is thrown when the desired queue position is not passed into
- *  TicketService.returnToQueue.
- *  @hidden */
-export const ERROR_NO_QUEUE_POSITION: string =
-  'Queue position missing from arguments.';
-/** @hidden */
-export const ERROR_INVALID_DESK: string =
-  'Desk is not a number or Desk object.';
-
-export type TicketCreationParameters = Pick<
-  Ticket,
-  'source' | 'firstName' | 'lastName' | 'phoneNumber' | 'email' | 'extra'
->;
-export type TicketEditingParameters = Pick<
-  Ticket,
-  'line' | 'phoneNumber' | 'firstName' | 'lastName' | 'email' | 'extra'
-> & { user: IdOrObject<User> };
-/**
- * The format of the HTTP request to send when creating a ticket.
- */
-interface TicketCreationRequest {
-  firstName?: string;
-  lastName?: string;
-  phoneNumber?: number;
-  email?: string;
-  extra?: string;
-  source?: string;
-}
-
-/**
- * The format of the HTTP request to send when editing a ticket.
- */
-interface TicketEditingRequest extends TicketCreationRequest {
-  line?: number;
-  user?: number;
-}
-
-interface TicketCallRequest {
-  user?: number | string;
-  desk?: number | string;
-  keepActiveTicketsOpen?: boolean;
-}
-
-interface CallNextRequest extends TicketCallRequest {
-  lines: string;
-}
-
-type TicketCreationResponse = Pick<Ticket, 'id'>;
+  search,
+  count,
+  create,
+  details,
+  edit,
+  call,
+  recall,
+  markServed,
+  markNoShow,
+  cancel,
+  returnToQueue,
+  addLabel,
+  setLabels,
+  removeLabel,
+  assignToUser,
+  unassign,
+  reorder,
+  getEstimatedTimeOfService,
+  getMessages,
+  sendMessage,
+  forward,
+  setExternalData,
+} from './ticket';
 
 /**
  * TicketService includes methods that work with tickets. Tickets represent a visitor that is
@@ -315,7 +66,7 @@ type TicketCreationResponse = Pick<Ticket, 'id'>;
  * // => { id: 141592145 }
  * ```
  */
-export class TicketService {
+export const TicketService = {
   /**
    * Searches for tickets according to the given search criteria.
    * Resolves to a list of tickets that match the search.
@@ -359,33 +110,7 @@ export class TicketService {
    * @param search the search criteria
    * @returns A promise that resolves to matching tickets.
    */
-  static search(search: TicketSearchCriteria): Promise<Array<Ticket>> {
-    const newSearch = { ...search };
-    if (newSearch.line && newSearch.line instanceof Array) {
-      newSearch.line = newSearch.line.join(',');
-    }
-    if (newSearch.status && newSearch.status instanceof Array) {
-      newSearch.status = newSearch.status.join(',');
-    }
-
-    if (newSearch.caller) {
-      newSearch.caller = extractIdToNumber(newSearch.caller);
-    } else {
-      delete newSearch.caller;
-    }
-
-    if (newSearch.responseScope && newSearch.responseScope instanceof Array) {
-      newSearch.responseScope = newSearch.responseScope.join(',');
-    }
-
-    const queryStr = new URLSearchParams(
-      newSearch as Record<string, string>,
-    ).toString();
-
-    return ApiBase.request(`tickets/search?${queryStr}`).then(
-      (response: { data: Ticket[] }) => response.data,
-    );
-  }
+  search,
 
   /**
    * Count all tickets that match the search criteria.
@@ -408,36 +133,7 @@ export class TicketService {
    * @param search the search criteria to use
    * @returns the number of tickets that match the search criteria
    */
-  static count(search: TicketCountCriteria): Promise<number> {
-    const newSearch = { ...search };
-    if (newSearch.line && newSearch.line instanceof Array) {
-      newSearch.line = newSearch.line.join(',');
-    }
-    if (newSearch.status && newSearch.status instanceof Array) {
-      newSearch.status = newSearch.status.join(',');
-    }
-    if (newSearch.caller) {
-      newSearch.caller = extractIdToNumber(newSearch.caller);
-    } else {
-      delete newSearch.caller;
-    }
-    // Sanity checks if user passes TicketSearchCriteria
-    if ((newSearch as TicketSearchCriteria).limit) {
-      delete (newSearch as TicketSearchCriteria).limit;
-    }
-    if ((newSearch as TicketSearchCriteria).order) {
-      delete (newSearch as TicketSearchCriteria).order;
-    }
-    if ((newSearch as TicketSearchCriteria).responseScope) {
-      delete (newSearch as TicketSearchCriteria).responseScope;
-    }
-    const queryStr = new URLSearchParams(
-      newSearch as Record<string, string>,
-    ).toString();
-    return ApiBase.request(`tickets/count?${queryStr}`).then(
-      (response: { count: number }) => response.count,
-    );
-  }
+  count,
 
   /**
    * Creates a new ticket and puts it into the queue as the last in the given line.
@@ -486,38 +182,7 @@ export class TicketService {
    * @returns a promise that resolves to the ID of the new ticket.
    * @throws ERROR_NO_LINE_ID when the lineId parameter is undefined or not a number.
    */
-  static create(
-    line: IdOrObject<Line>,
-    ticket: TicketCreationParameters,
-    idempotencyKey?: string | number,
-  ): Promise<TicketCreationResponse> {
-    if (line === undefined) {
-      throw new Error(ERROR_NO_LINE_ID);
-    }
-
-    const lineId = extractId(line);
-
-    const converted: any = { ...ticket };
-    if (converted.lastName === null) {
-      delete converted.lastName;
-    }
-    if (converted.extra) {
-      converted.extra = JSON.stringify(converted.extra);
-    }
-
-    const requestParams: TicketCreationRequest = { ...converted };
-
-    return ApiBase.request(
-      `lines/${lineId}/ticket`,
-      requestParams,
-      'POST',
-      idempotencyKey,
-    ).then((response: TicketCreationResponse) => {
-      const ticketId = parseInt(`${response.id}`, 10);
-      const reply: TicketCreationResponse = { id: ticketId };
-      return reply;
-    });
-  }
+  create,
 
   /**
    * Fetches the details of a given ticket ID and returns a Ticket object filled with data.
@@ -539,10 +204,7 @@ export class TicketService {
    * @returns the ticket's details as a Ticket object
    * @throws ERROR_NO_TICKET_ID when the ticket ID is undefined or not a number.
    */
-  static details(ticket: IdOrObject<Ticket>): Promise<Ticket> {
-    const ticketId = extractId(ticket);
-    return ApiBase.request(`tickets/${ticketId}`) as Promise<Ticket>;
-  }
+  details,
 
   /**
    * Edits the ticket.
@@ -565,34 +227,7 @@ export class TicketService {
    * @throws ERROR_NO_TICKET_ID when the ticket ID was undefined or not a number
    * @throws ERROR_NO_TICKET_CHANGES when the ticket changes were undefined
    */
-  static edit(
-    ticket: IdOrObject<Ticket>,
-    changes: TicketEditingParameters,
-  ): Promise<'success'> {
-    const ticketId = extractId(ticket);
-
-    if (!changes) {
-      throw new Error(ERROR_NO_TICKET_CHANGES);
-    }
-
-    const intermediate: any = { ...changes };
-
-    if (intermediate.extra) {
-      intermediate.extra = JSON.stringify(intermediate.extra);
-    }
-
-    if (intermediate.user) {
-      intermediate.user = extractId(intermediate.user);
-    } else {
-      delete intermediate.user;
-    }
-
-    const request: TicketEditingRequest = intermediate;
-
-    return ApiBase.request(`tickets/${ticketId}/edit`, request).then(
-      (response: { result: 'success' }) => response.result,
-    );
-  }
+  edit,
 
   /**
    * Call a specific ticket to service.
@@ -617,39 +252,7 @@ export class TicketService {
    * when calling a new ticket. This allows calling multiple tickets at the same time.
    * @returns  the ticket that was just called
    */
-  static call(
-    ticket: IdOrObject<Ticket>,
-    user?: IdOrObject<Ticket>,
-    desk?: IdOrObject<Ticket>,
-    keepActiveTicketsOpen?: boolean,
-  ): Promise<Ticket> {
-    if (!ticket) {
-      throw new Error(ERROR_NO_TICKET_ID);
-    }
-
-    const ticketId = extractId(ticket);
-
-    let request: TicketCallRequest = {};
-
-    if (user) {
-      request.user = extractId(user);
-    }
-
-    if (desk) {
-      request.desk = extractId(desk);
-    }
-
-    if (typeof keepActiveTicketsOpen !== 'undefined') {
-      request.keepActiveTicketsOpen = keepActiveTicketsOpen;
-    }
-
-    // If no request parameters specified, don't include the empty object.
-    if (!user && !desk && typeof keepActiveTicketsOpen === 'undefined') {
-      request = undefined;
-    }
-
-    return ApiBase.request(`tickets/${ticketId}/call`, request, 'POST');
-  }
+  call,
 
   /**
    * Recall a ticket.
@@ -661,18 +264,7 @@ export class TicketService {
    * @param ticket  The ticket to recall. The ticket ID can be used instead of the Ticket object.
    * @returns  a promise that resolves to 'success' if all went well.
    */
-  static recall(ticket: IdOrObject<Ticket>): Promise<'success'> {
-    const ticketId = extractId(ticket);
-
-    if (!ticketId || typeof ticketId !== 'string') {
-      throw new Error(ERROR_NO_TICKET_ID);
-    }
-    return ApiBase.request(
-      `tickets/${ticketId}/recall`,
-      undefined,
-      'POST',
-    ).then((response: { result: 'success' }) => response.result);
-  }
+  recall,
 
   /**
    * Mark a ticket served.
@@ -685,19 +277,7 @@ export class TicketService {
    * object.
    * @returns  a promise that resolves to 'success' if all went well.
    */
-  static markServed(ticket: IdOrObject<Ticket>): Promise<'success'> {
-    const ticketId = extractId(ticket);
-
-    if (!ticketId || typeof ticketId !== 'string') {
-      throw new Error(ERROR_NO_TICKET_ID);
-    }
-
-    return ApiBase.request(
-      `tickets/${ticketId}/markserved`,
-      undefined,
-      'POST',
-    ).then((response: { result: 'success' }) => response.result);
-  }
+  markServed,
 
   /**
    * Mark a ticket as "no-show".
@@ -711,18 +291,7 @@ export class TicketService {
    * @returns A promise that resolves to "success" when marking no-show works, and rejects when
    * something went wrong.
    */
-  static markNoShow(ticket: IdOrObject<Ticket>): Promise<'success'> {
-    const ticketId = extractId(ticket);
-
-    if (!ticketId || typeof ticketId !== 'string') {
-      throw new Error(ERROR_NO_TICKET_ID);
-    }
-    return ApiBase.request(
-      `tickets/${ticketId}/marknoshow`,
-      undefined,
-      'POST',
-    ).then((response: { result: 'success' }) => response.result);
-  }
+  markNoShow,
 
   /**
    * Cancel a ticket.
@@ -738,26 +307,7 @@ export class TicketService {
    * @returns  a promise that resolves to "success" if removing works, and rejects if something
    * went wrong.
    */
-  static cancel(
-    ticket: IdOrObject<Ticket>,
-    user: IdOrObject<User>,
-  ): Promise<string> {
-    const ticketId = extractId(ticket);
-    const userId = extractId(user);
-    if (!ticketId || typeof ticketId !== 'string') {
-      throw new Error(ERROR_NO_TICKET_ID);
-    }
-
-    if (!userId || typeof userId !== 'string') {
-      throw new Error(ERROR_NO_USER);
-    }
-
-    return ApiBase.request(
-      `tickets/${ticketId}/cancel`,
-      { user: userId },
-      'POST',
-    ).then((response: { result: 'success' }) => response.result);
-  }
+  cancel,
 
   /**
    * Return a ticket to the queue.
@@ -777,36 +327,7 @@ export class TicketService {
    * @returns {Promise<string>} a promise that resolves to "success" if it worked, and rejects
    * if something went wrong.
    */
-  static returnToQueue(
-    ticket: IdOrObject<Ticket>,
-    user: IdOrObject<User>,
-    position: DesiredQueuePosition,
-  ): Promise<'success'> {
-    const ticketId = extractId(ticket);
-    const userId = extractId(user);
-
-    if (!ticketId || typeof ticketId !== 'string') {
-      throw new Error(ERROR_NO_TICKET_ID);
-    }
-
-    if (!userId || typeof userId !== 'string') {
-      throw new Error('No user given');
-    }
-
-    if (!position) {
-      throw new Error(ERROR_NO_QUEUE_POSITION);
-    }
-
-    const query = new URLSearchParams({
-      position: `${position}`,
-      user: userId,
-    }).toString();
-    return ApiBase.request(
-      `tickets/${ticketId}/returntoqueue?${query}`,
-      undefined,
-      'POST',
-    ).then((response: { result: 'success' }) => response.result);
-  }
+  returnToQueue,
 
   /**
    * Add a label to a ticket.
@@ -832,34 +353,7 @@ export class TicketService {
    * @returns promise that resolves to 'success' if all was OK, and 'no
    * action' if the label was already there, and rejects if something else went wrong.
    */
-  static addLabel(
-    ticket: IdOrObject<Ticket>,
-    label: string,
-    user?: IdOrObject<User>,
-  ): Promise<'success' | 'no action'> {
-    const ticketId = extractId(ticket);
-    const userId = user ? extractId(user) : undefined;
-
-    if (!ticketId || typeof ticketId !== 'string') {
-      throw new Error(ERROR_NO_TICKET_ID);
-    }
-
-    if (!label) {
-      throw new Error('No label given.');
-    }
-
-    const body: { value: string; user?: string } = {
-      value: label,
-    };
-
-    if (typeof userId === 'string') {
-      body.user = userId;
-    }
-
-    return ApiBase.request(`tickets/${ticketId}/labels/add`, body, 'POST').then(
-      (response: { result: 'success' | 'no action' }) => response.result,
-    );
-  }
+  addLabel,
 
   /**
    * Set ticket labels.
@@ -883,28 +377,7 @@ export class TicketService {
    * @returns promise that resolves to 'success' if all was OK, and rejects
    * if something else went wrong.
    */
-  static setLabels(
-    ticket: IdOrObject<Ticket>,
-    labels: Array<string>,
-  ): Promise<'success'> {
-    const ticketId = extractId(ticket);
-
-    if (!ticketId || typeof ticketId !== 'string') {
-      throw new Error(ERROR_NO_TICKET_ID);
-    }
-
-    if (!labels) {
-      throw new Error('No labels given.');
-    }
-
-    const body: { labels: Array<string> } = { labels };
-
-    return ApiBase.request(
-      `tickets/${ticketId}/labels`,
-      JSON.stringify(body),
-      'PUT',
-    ).then((response: { result: 'success' }) => response.result);
-  }
+  setLabels,
 
   /**
    * Remove a label from the ticket.
@@ -931,37 +404,7 @@ export class TicketService {
    * @returns {Promise<string>}  A promise that resolves to "success" when removing the label
    * worked, and rejects when something went wrong.
    */
-  static removeLabel(
-    ticket: IdOrObject<Ticket>,
-    label: string,
-    user: IdOrObject<User>,
-  ): Promise<'success'> {
-    const ticketId = extractId(ticket);
-    const userId = extractId(user);
-
-    if (!ticketId || typeof ticketId !== 'string') {
-      throw new Error(ERROR_NO_TICKET_ID);
-    }
-
-    if (!label) {
-      throw new Error('No label given.');
-    }
-
-    if (!userId || typeof userId !== 'string') {
-      throw new Error('No user given');
-    }
-
-    const body = {
-      value: label,
-      user: userId,
-    };
-
-    return ApiBase.request(
-      `tickets/${ticketId}/labels/remove`,
-      body,
-      'POST',
-    ).then((response: { result: 'success' }) => response.result);
-  }
+  removeLabel,
 
   /**
    * Assign the given ticket to an user (assignee).
@@ -992,35 +435,7 @@ export class TicketService {
    * @param assignee The user who will take the ticket.
    * @returns {Promise.<string>} resolves to 'success' on success
    */
-  static assignToUser(
-    ticket: IdOrObject<Ticket>,
-    assigner: IdOrObject<User>,
-    assignee: IdOrObject<User>,
-  ): Promise<'success'> {
-    const ticketId = extractId(ticket);
-    const assignerId = extractId(assigner);
-    const assigneeId = extractId(assignee);
-
-    if (!ticketId || typeof ticketId !== 'string') {
-      throw new Error(ERROR_NO_TICKET_ID);
-    }
-
-    if (!assignerId || typeof assignerId !== 'string') {
-      throw new Error('No assigner given');
-    }
-
-    if (!assigneeId || typeof assigneeId !== 'string') {
-      throw new Error('No assignee given');
-    }
-
-    const body = {
-      assigner: assignerId,
-      assignee: assigneeId,
-    };
-    return ApiBase.request(`tickets/${ticketId}/assign`, body, 'POST').then(
-      (response: { result: 'success' }) => response.result,
-    );
-  }
+  assignToUser,
 
   /**
    * Un-assign a ticket. This returns the ticket to the unassigned list.
@@ -1063,29 +478,7 @@ export class TicketService {
    * @returns a Promise that resolves when unassigning works and rejects when
    * unassigning fails
    */
-  static unassign(
-    ticket: IdOrObject<Ticket>,
-    unassigner: IdOrObject<User>,
-  ): Promise<'success'> {
-    const ticketId = extractId(ticket);
-    const unassignerId = extractId(unassigner);
-
-    if (!ticketId || typeof ticketId !== 'string') {
-      throw new Error(ERROR_NO_TICKET_ID);
-    }
-
-    if (!unassignerId || typeof unassignerId !== 'string') {
-      throw new Error(
-        'Qminder.tickets.unassign was called without a valid unassigner user.',
-      );
-    }
-
-    return ApiBase.request(
-      `tickets/${ticketId}/unassign`,
-      { user: unassignerId },
-      'POST',
-    ).then((response: { result: 'success' }) => response.result);
-  }
+  unassign,
 
   /**
    * Reorder a ticket after another ticket.
@@ -1112,32 +505,7 @@ export class TicketService {
    * queue.
    * @returns resolves to 'success' when it worked
    */
-  static reorder(
-    ticket: IdOrObject<Ticket>,
-    afterTicket: IdOrObject<Ticket> | null,
-  ): Promise<'success'> {
-    const ticketId = extractId(ticket);
-
-    if (!ticketId || typeof ticketId !== 'string') {
-      throw new Error(ERROR_NO_TICKET_ID);
-    }
-
-    const afterTicketId: string | null =
-      afterTicket === null ? (afterTicket as null) : extractId(afterTicket);
-
-    let postData: { after: string | null } | undefined;
-    if (afterTicketId) {
-      postData = {
-        after: afterTicketId,
-      };
-    }
-
-    return ApiBase.request(
-      `tickets/${ticketId}/reorder`,
-      postData,
-      'POST',
-    ).then((response: { result: 'success' }) => response.result);
-  }
+  reorder,
 
   /**
    * Get the estimated time that this visitor will be called for service.
@@ -1160,20 +528,7 @@ export class TicketService {
    * of the Ticket object.
    * @returns the estimated Unix time the visitor will be called, eg 1509460809
    */
-  static getEstimatedTimeOfService(
-    ticket: IdOrObject<Ticket>,
-  ): Promise<number> {
-    const ticketId = extractId(ticket);
-
-    if (typeof ticketId !== 'string') {
-      throw new Error(ERROR_NO_TICKET_ID);
-    }
-
-    return ApiBase.request(`tickets/${ticketId}/estimated-time`).then(
-      (response: { estimatedTimeOfService: number }) =>
-        response.estimatedTimeOfService,
-    );
-  }
+  getEstimatedTimeOfService,
 
   /**
    * Get the ticket's SMS messages.
@@ -1206,18 +561,7 @@ export class TicketService {
    * @returns  a Promise that resolves to a list of ticket messages
    * @throws ERROR_NO_TICKET_ID  if the ticket is missing from the arguments, or invalid.
    */
-  static getMessages(
-    ticket: IdOrObject<Ticket>,
-  ): Promise<Array<TicketMessage>> {
-    const ticketId = extractId(ticket);
-
-    if (!ticketId || typeof ticketId !== 'string') {
-      throw new Error(ERROR_NO_TICKET_ID);
-    }
-    return ApiBase.request(`tickets/${ticketId}/messages`).then(
-      (response: { messages: TicketMessage[] }) => response.messages,
-    );
-  }
+  getMessages,
 
   /**
    * Send a new SMS message to a visitor.
@@ -1241,9 +585,9 @@ export class TicketService {
    *        "Hello! Free coffee time!",
    *        { id: 14245 }
    * ).then(function(success) {
-   *     console.log("It works!");
+   *     console.log('It works!');
    * }, function(error) {
-   *     console.log("Something went wrong while sending the message.");
+   *     console.log('Something went wrong while sending the message.');
    * });
    * ```
    * @param ticket  The ticket to send a message to. The user ID may also be used.
@@ -1252,35 +596,7 @@ export class TicketService {
    * @returns a promise that resolves to the string "success" if it works, and rejects when
    * something goes wrong.
    */
-  static sendMessage(
-    ticket: IdOrObject<Ticket>,
-    message: string,
-    user: IdOrObject<User>,
-  ): Promise<'success'> {
-    const ticketId = extractId(ticket);
-    const userId = extractId(user);
-
-    if (!ticketId || typeof ticketId !== 'string') {
-      throw new Error(ERROR_NO_TICKET_ID);
-    }
-
-    if (!message || typeof message !== 'string') {
-      throw new Error('No message specified. The message has to be a string.');
-    }
-
-    if (!userId || typeof userId !== 'string') {
-      throw new Error(ERROR_NO_USER);
-    }
-
-    const body = {
-      message,
-      user: userId,
-    };
-
-    return ApiBase.request(`tickets/${ticketId}/messages`, body, 'POST').then(
-      (response: { result: 'success' }) => response.result,
-    );
-  }
+  sendMessage,
 
   /**
    * Forward the ticket to another queue.
@@ -1320,38 +636,7 @@ export class TicketService {
    * @returns  a Promise that resolves when forwarding works, and rejects when it fails.
    * @throws an Error when the ticket or line are missing or invalid.
    */
-  static forward(
-    ticket: IdOrObject<Ticket>,
-    line: IdOrObject<Line>,
-    user?: IdOrObject<User>,
-  ): Promise<object> {
-    const ticketId = extractId(ticket);
-    const lineId = extractId(line);
-    const userId = extractId(user);
-
-    if (typeof ticketId !== 'string') {
-      throw new Error(ERROR_NO_TICKET_ID);
-    }
-
-    if (typeof lineId !== 'string') {
-      throw new Error('Line ID is missing.');
-    }
-
-    // If the user's ID was passed and it's invalid, throw an error
-    if (user !== undefined && typeof userId !== 'string') {
-      throw new Error('User ID is missing.');
-    }
-
-    const body: { line: string; user?: string } = {
-      line: lineId,
-    };
-
-    if (userId !== undefined) {
-      body.user = userId;
-    }
-
-    return ApiBase.request(`tickets/${ticketId}/forward`, body);
-  }
+  forward,
 
   /**
    * Add external data to a ticket.
@@ -1362,42 +647,5 @@ export class TicketService {
    * @param data      The data to set
    * @returns promise that resolves to 'success' if all was OK and rejects if something else went wrong.
    */
-  static setExternalData(
-    ticket: IdOrObject<Ticket>,
-    provider: string,
-    title: string,
-    data: any,
-  ): Promise<'success'> {
-    const ticketId = extractId(ticket);
-
-    if (!ticketId || typeof ticketId !== 'string') {
-      throw new Error(ERROR_NO_TICKET_ID);
-    }
-
-    if (!provider || typeof provider !== 'string') {
-      throw new Error(
-        'No provider specified. The provider has to be a string.',
-      );
-    }
-
-    if (!title || typeof title !== 'string') {
-      throw new Error('No title specified. The title has to be a string.');
-    }
-
-    if (!data) {
-      throw new Error('No data provided.');
-    }
-
-    const payload = {
-      provider,
-      title,
-      data: JSON.stringify(data),
-    };
-
-    return ApiBase.request(
-      `tickets/${ticketId}/external`,
-      payload,
-      'POST',
-    ).then((response: { result: 'success' }) => response.result);
-  }
-}
+  setExternalData,
+};
